# backend/internal/api/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn5AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn1AllowedHeaders = map[string]string{
		"GET":  "Authorization",
		"POST": "Authorization,Content-Type",
	}
	rn3AllowedHeaders = map[string]string{
		"DELETE": "Authorization",
		"GET":    "Authorization",
		"PUT":    "Authorization,Content-Type",
	}
	rn7AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn9AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn10AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'r': // Prefix: "reco"

				if l := len("reco"); len(elem) >= l && elem[0:l] == "reco" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'm': // Prefix: "mmend"

					if l := len("mmend"); len(elem) >= l && elem[0:l] == "mmend" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleGetRecommendRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: rn5AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				case 'r': // Prefix: "rds"

					if l := len("rds"); len(elem) >= l && elem[0:l] == "rds" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch r.Method {
						case "GET":
							s.handleListRecordsRequest([0]string{}, elemIsEscaped, w, r)
						case "POST":
							s.handleCreateRecordRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET,POST",
								allowedHeaders: rn1AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						// Param: "id"
						// Leaf parameter, slashes are prohibited
						idx := strings.IndexByte(elem, '/')
						if idx >= 0 {
							break
						}
						args[0] = elem
						elem = ""

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "DELETE":
								s.handleDeleteRecordRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							case "GET":
								s.handleGetRecordRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							case "PUT":
								s.handleUpdateRecordRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "DELETE,GET,PUT",
									allowedHeaders: rn3AllowedHeaders,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

					}

				}

			case 's': // Prefix: "stats/"

				if l := len("stats/"); len(elem) >= l && elem[0:l] == "stats/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'f': // Prefix: "flavor-words"

					if l := len("flavor-words"); len(elem) >= l && elem[0:l] == "flavor-words" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleGetStatsFlavorWordsRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: rn7AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				case 's': // Prefix: "summary"

					if l := len("summary"); len(elem) >= l && elem[0:l] == "summary" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleGetStatsSummaryRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET",
								allowedHeaders: rn9AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			case 'u': // Prefix: "userinfo"

				if l := len("userinfo"); len(elem) >= l && elem[0:l] == "userinfo" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleGetUserinfoRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: rn10AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'r': // Prefix: "reco"

				if l := len("reco"); len(elem) >= l && elem[0:l] == "reco" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'm': // Prefix: "mmend"

					if l := len("mmend"); len(elem) >= l && elem[0:l] == "mmend" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = GetRecommendOperation
							r.summary = "Get recommendation score for a bean"
							r.operationID = "getRecommend"
							r.operationGroup = ""
							r.pathPattern = "/recommend"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 'r': // Prefix: "rds"

					if l := len("rds"); len(elem) >= l && elem[0:l] == "rds" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch method {
						case "GET":
							r.name = ListRecordsOperation
							r.summary = "List records"
							r.operationID = "listRecords"
							r.operationGroup = ""
							r.pathPattern = "/records"
							r.args = args
							r.count = 0
							return r, true
						case "POST":
							r.name = CreateRecordOperation
							r.summary = "Create a new record"
							r.operationID = "createRecord"
							r.operationGroup = ""
							r.pathPattern = "/records"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						// Param: "id"
						// Leaf parameter, slashes are prohibited
						idx := strings.IndexByte(elem, '/')
						if idx >= 0 {
							break
						}
						args[0] = elem
						elem = ""

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "DELETE":
								r.name = DeleteRecordOperation
								r.summary = "Delete a record"
								r.operationID = "deleteRecord"
								r.operationGroup = ""
								r.pathPattern = "/records/{id}"
								r.args = args
								r.count = 1
								return r, true
							case "GET":
								r.name = GetRecordOperation
								r.summary = "Get a record by ID"
								r.operationID = "getRecord"
								r.operationGroup = ""
								r.pathPattern = "/records/{id}"
								r.args = args
								r.count = 1
								return r, true
							case "PUT":
								r.name = UpdateRecordOperation
								r.summary = "Update a record"
								r.operationID = "updateRecord"
								r.operationGroup = ""
								r.pathPattern = "/records/{id}"
								r.args = args
								r.count = 1
								return r, true
							default:
								return
							}
						}

					}

				}

			case 's': // Prefix: "stats/"

				if l := len("stats/"); len(elem) >= l && elem[0:l] == "stats/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'f': // Prefix: "flavor-words"

					if l := len("flavor-words"); len(elem) >= l && elem[0:l] == "flavor-words" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = GetStatsFlavorWordsOperation
							r.summary = "Get flavor word frequency"
							r.operationID = "getStatsFlavorWords"
							r.operationGroup = ""
							r.pathPattern = "/stats/flavor-words"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 's': // Prefix: "summary"

					if l := len("summary"); len(elem) >= l && elem[0:l] == "summary" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = GetStatsSummaryOperation
							r.summary = "Get statistics summary"
							r.operationID = "getStatsSummary"
							r.operationGroup = ""
							r.pathPattern = "/stats/summary"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			case 'u': // Prefix: "userinfo"

				if l := len("userinfo"); len(elem) >= l && elem[0:l] == "userinfo" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = GetUserinfoOperation
						r.summary = "Get current user info (proxied from OIDC provider, or dummy in disable-oidc mode)"
						r.operationID = "getUserinfo"
						r.operationGroup = ""
						r.pathPattern = "/userinfo"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			}

		}
	}
	return r, false
}