# openapi.yaml
openapi: "3.0.3"
info:
  title: beanmemo API
  description: Personal coffee bean diary API
  version: "1.0.0"

servers:
  - url: /api

security:
  - bearerAuth: []

paths:
  /records:
    get:
      operationId: listRecords
      summary: List records
      parameters:
        - name: origin
          in: query
          schema:
            type: string
        - name: roast_level
          in: query
          schema:
            type: string
            enum: [light, medium, dark]
        - name: rating_min
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: brew_method
          in: query
          schema:
            type: string
            enum: [drip, espresso, french_press, other]
      responses:
        "200":
          description: List of records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Record"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createRecord
      summary: Create a new record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecordRequest"
      responses:
        "201":
          description: Created record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Record"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /records/{id}:
    get:
      operationId: getRecord
      summary: Get a record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Record detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateRecord
      summary: Update a record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecordRequest"
      responses:
        "200":
          description: Updated record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Record"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteRecord
      summary: Delete a record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /userinfo:
    get:
      operationId: getUserinfo
      summary: Get current user info (proxied from OIDC provider, or dummy in disable-oidc mode)
      responses:
        "200":
          description: User info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserinfoResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /stats/summary:
    get:
      operationId: getStatsSummary
      summary: Get statistics summary
      responses:
        "200":
          description: Statistics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /stats/flavor-words:
    get:
      operationId: getStatsFlavorWords
      summary: Get flavor word frequency
      responses:
        "200":
          description: Flavor words with counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlavorWord"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /recommend:
    get:
      operationId: getRecommend
      summary: Get recommendation score for a bean
      parameters:
        - name: origin
          in: query
          schema:
            type: string
        - name: name
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Recommendation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Record:
      type: object
      required:
        - id
        - name
        - rating
        - is_note_filled
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        origin:
          type: string
          nullable: true
        roast_level:
          type: string
          nullable: true
          enum: [light, medium, dark]
        shop:
          type: string
          nullable: true
        price:
          type: integer
          nullable: true
        purchased_at:
          type: string
          format: date
          nullable: true
        tasting_note:
          type: string
          nullable: true
        brew_method:
          type: string
          nullable: true
          enum: [drip, espresso, french_press, other]
        recipe:
          type: string
          nullable: true
        is_note_filled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    RecordDetail:
      allOf:
        - $ref: "#/components/schemas/Record"
        - type: object
          required:
            - related_records
          properties:
            related_records:
              type: array
              items:
                $ref: "#/components/schemas/Record"
              description: Other records of the same bean (same name or origin)

    CreateRecordRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
        rating:
          type: integer
          minimum: 1
          maximum: 5
        origin:
          type: string
          nullable: true
        roast_level:
          type: string
          nullable: true
          enum: [light, medium, dark]
        shop:
          type: string
          nullable: true
        price:
          type: integer
          nullable: true
          minimum: 0
        purchased_at:
          type: string
          format: date
          nullable: true
        tasting_note:
          type: string
          nullable: true
        brew_method:
          type: string
          nullable: true
          enum: [drip, espresso, french_press, other]
        recipe:
          type: string
          nullable: true

    UpdateRecordRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        rating:
          type: integer
          minimum: 1
          maximum: 5
        origin:
          type: string
          nullable: true
        roast_level:
          type: string
          nullable: true
          enum: [light, medium, dark]
        shop:
          type: string
          nullable: true
        price:
          type: integer
          nullable: true
          minimum: 0
        purchased_at:
          type: string
          format: date
          nullable: true
        tasting_note:
          type: string
          nullable: true
        brew_method:
          type: string
          nullable: true
          enum: [drip, espresso, french_press, other]
        recipe:
          type: string
          nullable: true

    StatsSummary:
      type: object
      required:
        - by_origin
        - by_roast_level
        - by_brew_method
        - total_records
      properties:
        total_records:
          type: integer
        by_origin:
          type: array
          items:
            $ref: "#/components/schemas/GroupStat"
        by_roast_level:
          type: array
          items:
            $ref: "#/components/schemas/GroupStat"
        by_brew_method:
          type: array
          items:
            $ref: "#/components/schemas/GroupStat"

    GroupStat:
      type: object
      required:
        - label
        - count
        - avg_rating
      properties:
        label:
          type: string
        count:
          type: integer
        avg_rating:
          type: number
          format: float

    FlavorWord:
      type: object
      required:
        - word
        - count
      properties:
        word:
          type: string
        count:
          type: integer

    RecommendResult:
      type: object
      required:
        - locked
        - total_records
      properties:
        locked:
          type: boolean
        records_needed:
          type: integer
          nullable: true
          description: How many more records needed to unlock (only when locked=true)
        score:
          type: number
          format: float
          nullable: true
          description: Preference score 0.0-5.0 (null when locked)
        total_records:
          type: integer
        origin_avg:
          type: number
          format: float
          nullable: true
        name_match_avg:
          type: number
          format: float
          nullable: true

    UserinfoResponse:
      type: object
      required:
        - sub
      properties:
        sub:
          type: string
        name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        picture:
          type: string
          nullable: true

    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string